+++ source/lue_cxx/include/lue/property.hpp
@@ -5,6 +5,9 @@
 
 namespace lue {
 
+/*!
+    @brief      TODO
+*/
 class Property:
     public hdf5::Group
 {