+++ source/lue_cxx/include/lue/value.hpp
@@ -4,6 +4,9 @@
 
 namespace lue {
 
+/*!
+    @brief      TODO
+*/
 class Value
 {