+++ source/lue_py/test/constant_size/time/omnipresent/discretization_test.py
@@ -17,7 +17,7 @@ class DiscretizationTest(lue_test.TestCase):
 
         self.assertEqual(property_set.name, name)
         self.assertEqual(type(property_set), omnipresent.PropertySet)
-        self.assertArraysEqual(property_set.domain.space.boxes[:], boxes)
+        self.assertArraysEqual(property_set.domain.space.items[:], boxes)
 
 
     def verify_property_values(self,